Consolidate JS and TS configs
