+++ fixtures/component.js
