+++ fixtures/import.ts
@@ -2,4 +2,4 @@ import something, {foo, baz} from './exports' // baz is not exported
 import somethingElse from './nonexistent' // file does not exist
 
 export {foo, baz, something, somethingElse}
-export {bar} from '../ts/exports'
+export {bar} from './exports'
