+++ fixtures/variables.js
