+++ fixtures/variables.ts
